+++ packages/table/src/Table.tsx
@@ -213,6 +213,12 @@ function TableRender<T extends Record<string, any>, U, ValueType>(
       )
     : baseTableDom;
 
+  useEffect(() => {
+    if (props.name) {
+      counter.setEditorTableForm(props.editable!.form!);
+    }
+  }, [counter, props.editable, props.name]);
+
   const tableContentDom = useMemo(() => {
     if (props.editable && !props.name) {
       return (
@@ -237,9 +243,11 @@ function TableRender<T extends Record<string, any>, U, ValueType>(
             dateFormatter={props.dateFormatter}
             contentRender={(items: React.ReactNode) => {
               if (counter.editableForm) return items;
+              if (props.loading === false) return;
+              const loadingProps = props.loading === true ? {} : props.loading;
               return (
                 <div style={{ paddingTop: 100, textAlign: 'center' }}>
-                  <Spin size="large" />
+                  <Spin size="large" {...loadingProps} />
                 </div>
               );
             }}
@@ -258,7 +266,7 @@ function TableRender<T extends Record<string, any>, U, ValueType>(
       </>
     );
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [alertDom, !!props.editable, tableDom, toolbarDom]);
+  }, [alertDom, props.loading, !!props.editable, tableDom, toolbarDom]);
 
   /** Table 区域的 dom，为了方便 render */
   const tableAreaDom =
@@ -802,23 +810,17 @@ const ProviderWarp = <
   props: ProTableProps<DataType, Params, ValueType>,
 ) => {
   const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
-  const ErrorComponent = props.ErrorBoundary || ErrorBoundary;
+  const ErrorComponent =
+    props.ErrorBoundary === false ? React.Fragment : props.ErrorBoundary || ErrorBoundary;
   return (
     <Container.Provider initialState={props}>
       <ConfigProviderWrap>
-        {props.ErrorBoundary === false ? (
+        <ErrorComponent>
           <ProTable<DataType, Params, ValueType>
             defaultClassName={getPrefixCls('pro-table')}
             {...props}
           />
-        ) : (
-          <ErrorComponent>
-            <ProTable<DataType, Params, ValueType>
-              defaultClassName={getPrefixCls('pro-table')}
-              {...props}
-            />
-          </ErrorComponent>
-        )}
+        </ErrorComponent>
       </ConfigProviderWrap>
     </Container.Provider>
   );