+++ scripts/release.js
@@ -142,8 +142,9 @@ async function release() {
       console.log(
         `[${index + 1}/${pkgs.length}] Publish package ${name} ${isNext ? 'with next tag' : ''}`,
       );
-      const cliArgs = isNext ? ['publish', '--tag', 'next'] : ['publish'];
-      const { stdout } = execa.sync('npm', cliArgs, {
+      // 默认设置为 tag 检查通过之后在设置为 latest
+      const cliArgs = isNext ? ['publish', '--tag', 'next'] : ['publish', '--tag', 'beta'];
+      const { stdout } = execa('npm', cliArgs, {
         cwd: pkgPath,
       });
       console.log(stdout);